moved the test files to the same package as implementation files to be able to calculate coverage.
